Hocus Focus Committed
This game is compitable for samsung mobiles with Android 2.3.3 &
above.
The game is inspired by a popular Tele kids NewsPaper game "Hocus
Focus". The game is focused in improving the memory and image
recognition capabilities of brain.
Objective of the game is to find and
mark the dissimilarities between very similar images.